#393 [refactor] controller 아래 common 폴더 생성
