+++ module-api/src/main/java/com/mile/common/resolver/comment/CommentVariableResolver.java
@@ -1,4 +1,4 @@
-package com.mile.resolver.moim;
+package com.mile.common.resolver.comment;
 
 import com.mile.exception.message.ErrorMessage;
 import com.mile.exception.model.BadRequestException;
@@ -18,13 +18,13 @@ import java.util.Map;
 
 @Component
 @RequiredArgsConstructor
-public class MoimVariableResolver implements HandlerMethodArgumentResolver {
-    private static final String MOIM_PATH_VARIABLE = "moimId";
+public class CommentVariableResolver implements HandlerMethodArgumentResolver {
+    private static final String COMMENT_PATH_VARIABLE = "commentId";
     private final SecureUrlUtil secureUrlUtil;
 
     @Override
     public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.hasParameterAnnotation(MoimIdPathVariable.class);
+        return parameter.hasParameterAnnotation(CommentIdPathVariable.class);
     }
 
     @Override
@@ -32,9 +32,9 @@ public class MoimVariableResolver implements HandlerMethodArgumentResolver {
         final HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
         final Map<String, String> pathVariables = (Map<String, String>) request.getAttribute(HandlerMapping.URI_TEMPLATE_VARIABLES_ATTRIBUTE);
 
-        final String moimId = pathVariables.get(MOIM_PATH_VARIABLE);
+        final String id = pathVariables.get(COMMENT_PATH_VARIABLE);
         try {
-            return secureUrlUtil.decodeUrl(moimId);
+            return secureUrlUtil.decodeUrl(id);
         } catch (NumberFormatException e) {
             throw new BadRequestException(ErrorMessage.INVALID_URL_EXCEPTION);
         }
