+++ module-api/src/main/java/com/mile/common/resolver/reply/ReplyIdPathVariable.java
@@ -1,4 +1,4 @@
-package com.mile.resolver.moim;
+package com.mile.common.resolver.reply;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,6 +7,5 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface MoimIdPathVariable {
-
+public @interface ReplyIdPathVariable {
 }
