+++ module-api/src/main/java/com/mile/common/resolver/topic/TopicIdPathVariable.java
@@ -1,4 +1,4 @@
-package com.mile.resolver.post;
+package com.mile.common.resolver.topic;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,5 +7,6 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.PARAMETER)
 @Retention(RetentionPolicy.RUNTIME)
-public @interface PostIdPathVariable {
+public @interface TopicIdPathVariable {
+
 }
